requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
PyMySQL==1.0.3
Flask-Migrate==4.0.4
python-dotenv==1.0.0
flask_wtf==1.2.2
flask_login==0.6.3